changing the ad pool in dropdown component
