+++ src/app/campaigns/components/ad-pool-token-select/ad-pool-token-select.component.html
@@ -1,4 +1,4 @@
-<div class="modal_container">
+<!-- <div class="modal_container">
   <h3 class="modal_title w-fit mx-auto">Select token</h3>
   <p class="modal_subtitle">
     YOUR NETWORK:
@@ -17,7 +17,7 @@
   </div>
   <div class="container">
     <!-- Your content for each crypto element here -->
-    <div class="row">
+<!-- <div class="row">
       <div
         class="col-sm-4"
         *ngFor="let crypto of filterList.slice(0, 6); let i = index"
@@ -28,8 +28,8 @@
         </div>
       </div>
     </div>
-  </div>
-  <hr />
+  </div> -->
+<!-- <hr />
   <div>
     <ng-container *ngIf="tokenNotFound && showWarning">
       <div class="warning_div">
@@ -92,13 +92,13 @@
         >
           <img class="img_token" src="{{ crypto.value.logo }}" alt="Logo" />
           <!-- <ul *ngFor='let item of tokens'></ul> -->
-          <p class="name_token p-2">{{ crypto.value.name }}</p>
-          <p class="balance_token ml-auto p-2">0</p>
-          <!-- justify-content-end -->
-        </div>
-      </div>
-    </ng-container>
-  </div>
+<!-- <p class="name_token p-2">{{ crypto.value.name }}</p> -->
+<!-- <p class="balance_token ml-auto p-2">0</p> -->
+<!-- justify-content-end -->
+<!-- </div> -->
+<!-- </div> -->
+<!-- </ng-container> -->
+<!-- </div>  -->
 
-  <!-- the token list will appear here  with an overflow cause there is more than 100 token-->
-</div>
+<!-- the token list will appear here  with an overflow cause there is more than 100 token-->
+<!-- </div> -->
