+++ src/app/campaigns/components/ad-pool-token-select/ad-pool-token-select.component.spec.ts
@@ -8,9 +8,8 @@ describe('AdPoolTokenSelectComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ AdPoolTokenSelectComponent ]
-    })
-    .compileComponents();
+      declarations: [AdPoolTokenSelectComponent]
+    }).compileComponents();
   });
 
   beforeEach(() => {
