+++ src/app/campaigns/components/remuneration/remuneration.component.html
@@ -14,17 +14,15 @@
           'create_campaign_form.budjet' | translate
         }}</label>
       </div>
-      <ng-template #tokenModal>
+      <!-- <ng-template #tokenModal>
         <app-ad-pool-token-select
-          (click)="closeTokenModal(tokenModal)"
-          [closeModal]="openModalToken"
           (tokenSelected)="onCryptoSelected($event)"
           [selectedNetworkValue]="selectedCryptoDetails.network"
         ></app-ad-pool-token-select>
-      </ng-template>
-      <button (click)="createTokenModal()">
-        Click Me : {{ selectedToken?.name }}
-      </button>
+      </ng-template> -->
+      <!-- <button (click)="createTokenModal()">
+        Click Me : {{ selectedToken?.value.name }}
+      </button> -->
 
       <div class="height-drop">
         <app-dropdown-crypto-network
