+++ src/app/wallet/components/dropdown-crypto-network/dropdown-crypto-network.component.html
@@ -66,6 +66,7 @@
     >
       <button
         [disabled]="selectedNetworkValue === 'BTC'"
+        (click)="selectToken(selectTokenModal)"
         class="btn btntoggle dropdown-toggle btn-dropdown"
       >
         <i class="chevron fa fa-fw" *ngIf="selectedNetworkValue !== 'BTC'"></i>
@@ -90,6 +91,7 @@
         class="dropdown-menu"
         aria-labelledby="dropdownButton2"
         id="dropdownButtonUl"
+        *ngIf="!router.url.startsWith('/campaign')"
       >
         <ng-container *ngFor="let crypto of dataList; trackBy: trackById">
           <li
@@ -127,3 +129,110 @@
     </div>
   </div>
 </div>
+
+<ng-template #selectTokenModal>
+  <div class="modal_container">
+    <h3 class="modal_title w-fit mx-auto">Select token</h3>
+    <p class="modal_subtitle">
+      YOUR NETWORK:
+      <span class="network_selected">{{ selectedNetworkValue }}</span>
+    </p>
+    <div class="input_search_token_container">
+      <i id="search_logo" class="bi bi-search"></i>
+      <input
+        (click)="reset($event)"
+        class="input_search_token"
+        (input)="searchToken($event)"
+        type="text"
+        [formControl]="tokenSearch"
+        placeholder="Search a name or paste address"
+      />
+    </div>
+    <div class="container">
+      <!-- Your content for each crypto element here -->
+      <div class="row">
+        <div
+          class="col-sm-4"
+          *ngFor="let crypto of filterList.slice(0, 6); let i = index"
+        >
+          <div class="tokens_preview d-flex p-2 mb-2">
+            <img class="img_token" src="{{ crypto.logo }}" alt="logo" />
+            <p class="name_token">{{ crypto.key }}</p>
+          </div>
+        </div>
+      </div>
+    </div>
+    <hr />
+    <div>
+      <ng-container *ngIf="tokenNotFound && showWarning">
+        <div class="warning_div">
+          <p>{{ 'No Search Result Available' | uppercase }}</p>
+          <p
+            class="warning_message"
+            *ngIf="!showSearchNewTokenContainer"
+            (click)="searchPersonalizedToken()"
+          >
+            Add a personnaliszed token >
+          </p>
+        </div>
+      </ng-container>
+      <ng-container *ngIf="showSearchNewTokenContainer">
+        <div class="form-group">
+          <label class="label_title" for="contractAddress">
+            {{ "Token's Contract Address" | uppercase }}
+          </label>
+          <input
+            id="contractAddress"
+            class="input_search_token_container"
+            type="text"
+            placeholder="0x"
+          />
+        </div>
+
+        <div class="form-group">
+          <label class="label_title" for="tokenSymbol">
+            {{ "Token's Symbol" | uppercase }}
+          </label>
+          <input
+            id="tokenSymbol"
+            class="input_search_token_container"
+            type="text"
+            placeholder="SOL"
+          />
+        </div>
+
+        <div class="form-group">
+          <label class="label_title" for="decimals">
+            {{ 'Decimals' | uppercase }}
+          </label>
+          <input
+            id="decimals"
+            class="input_search_token_container"
+            type="text"
+            placeholder="18"
+          />
+        </div>
+
+        <button class="button_add_token">+ Add Token</button>
+      </ng-container>
+      <ng-container *ngIf="!tokenNotFound">
+        <div class="list_crypto_container">
+          <p class="modal_subtitle">Token List</p>
+          <div
+            *ngFor="let crypto of filterList"
+            (click)="tokenToSelect(crypto)"
+            class="list_token d-flex"
+          >
+            <img class="img_token" src="{{ crypto.logo }}" alt="Logo" />
+            <!-- <ul *ngFor='let item of tokens'></ul> -->
+            <p class="name_token p-2">{{ crypto.name }}</p>
+            <p class="balance_token ml-auto p-2">0</p>
+            <!-- justify-content-end -->
+          </div>
+        </div>
+      </ng-container>
+    </div>
+
+    <!-- the token list will appear here  with an overflow cause there is more than 100 token-->
+  </div>
+</ng-template>
