+++ src/app/wallet/components/dropdown-crypto-network/dropdown-crypto-network.component.ts
@@ -7,11 +7,16 @@ import {
   OnDestroy,
   OnInit,
   Output,
-  SimpleChanges
+  SimpleChanges,
+  TemplateRef,
+  ViewChild
 } from '@angular/core';
+import { FormControl } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ListTokens } from '@app/config/atn.config';
 import { WalletFacadeService } from '@app/core/facades/wallet-facade.service';
+import { CryptofetchServiceService } from '@app/core/services/wallet/cryptofetch-service.service';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { Subject } from 'rxjs';
 import { filter, takeUntil } from 'rxjs/operators';
 
@@ -24,16 +29,30 @@ export class DropdownCryptoNetworkComponent
   implements OnInit, OnDestroy, OnChanges
 {
   @Output() selectedCrypto: EventEmitter<any> = new EventEmitter();
+  @ViewChild('tokenModal', { static: false })
+  private tokenModal!: TemplateRef<any>;
   selectedNetworkValue: string = 'ERC20';
   cryptoPicName: string = 'SATT';
   cryptoSymbol: string = 'SATT';
   cryptoList$ = this.walletFacade.cryptoList$;
+  cryptoListApi: any = [];
   dataList: any = [];
   defaultcurr: any;
   defaultcurrbtt: any;
+  tokenSearch = new FormControl('');
+  // @Input() selectedToken: any = [];
+  @Output() tokenSelected = new EventEmitter<any>();
+  campaignCryptoList: any = [];
+  filterList: any = [];
+  userCrypto: any = [];
+
+  tokenNotFound: boolean = false;
+  showWarning: boolean = true;
+  showSearchNewTokenContainer: boolean = false;
 
   defaultcurrbep: any;
   defaultcurrbtc: any;
+  selectedToken: any;
   cryptoName: any;
   token: any;
   networkList: Array<{ network: string }>;
@@ -51,11 +70,65 @@ export class DropdownCryptoNetworkComponent
   cryptoList: any = [];
   defaultcurrpolygon: any;
   defaultcurrtron: any;
+
+  @ViewChild('selectToken', { static: false })
+  public selectTokenModal!: TemplateRef<any>;
+  tokenList: any = [];
+
+  reset(e: any) {
+    e.target.value = '';
+    this.filterList = this.cryptoList;
+    this.showSearchNewTokenContainer = false;
+    this.showWarning = true;
+    this.tokenNotFound = false;
+  }
+
+  searchToken(e: any) {
+    this.filterList = [];
+
+    if (e.target.value.length > 0) {
+      this.campaignCryptoList.forEach((crypto: any) => {
+        if (
+          crypto.value.name
+            .toString()
+            .toLowerCase()
+            .includes(e.target.value.toLowerCase()) // crypto.symbol.includes(e.target.value)
+        )
+          this.filterList.push(crypto);
+        console.log(this.filterList), 'filterList';
+      });
+      if (this.filterList.length === 0) this.tokenNotFound = true;
+      else this.tokenNotFound = false;
+    } else this.filterList = this.campaignCryptoList;
+  }
+
+  searchPersonalizedToken() {
+    this.showSearchNewTokenContainer = !this.showSearchNewTokenContainer;
+    this.showWarning = false;
+  }
+
+  tokenToSelect(crypto: any) {
+    console.log(crypto, 'crypto selected');
+    // this.getTokenBalance();
+    this.tokenSelected.emit(crypto);
+  }
+
+  getUserCrypto() {
+    this.walletFacade.cryptoList$.subscribe((data: any) => {
+      data = JSON.parse(JSON.stringify(data));
+      this.userCrypto = data;
+      console.log(this.userCrypto, 'userCrypto');
+    });
+  }
+
   constructor(
+    public modalService: NgbModal,
     private walletFacade: WalletFacadeService,
     private route: ActivatedRoute,
     public router: Router,
-    private cdref: ChangeDetectorRef
+    private cdref: ChangeDetectorRef,
+
+    private Fetchservice: CryptofetchServiceService
   ) {
     this.networkList = [
       { network: 'BEP20' },
@@ -65,6 +138,20 @@ export class DropdownCryptoNetworkComponent
       { network: 'BTC' },
       { network: 'TRON' }
     ];
+  }
+
+  createTokenModal() {
+    // this.walletFacade.getCryptoPriceList().subscribe((res) => {
+    //   console.log(res);
+    // });
+
+    if (this.router.url.startsWith('/campaign')) {
+      this.modalService.open(this.tokenModal);
+    }
+  }
+
+  closeTokenModal(content: any) {
+    this.modalService.dismissAll(content);
   }
 
   ngOnInit(): void {
@@ -97,6 +184,57 @@ export class DropdownCryptoNetworkComponent
     this.defaultcurrpolygon = ListTokens['MATIC'].name;
     this.defaultcurrbtt = ListTokens['BTT'].name;
     this.defaultcurrtron = ListTokens['TRX'].name;
+
+    //SELECT TOKEN FOR CREATE CAMPAGIN
+
+    if (this.router.url.startsWith('/campaign')) {
+      this.walletFacade.getCryptoPriceList().subscribe((res: any) => {
+        console.log({ data: res.data }, 'data wallet facade');
+        const result = Object.keys(res.data);
+        result.forEach((key) => {
+          let arr = res?.data[key]?.networkSupported || [];
+          if (!res.data[key].network) {
+            arr.forEach((data: any) => {
+              data.platform?.name
+                .toUpperCase()
+                .includes(this.selectedNetworkValue) &&
+                !this.campaignCryptoList.find(
+                  (e: any) => e.name === res.data[key].name
+                ) &&
+                this.campaignCryptoList.push({
+                  key: key,
+                  value: res.data[key]
+                });
+            });
+          } else {
+            res.data[key].network === this.selectedNetworkValue &&
+              this.campaignCryptoList.push(res.data[key]);
+          }
+        });
+        this.filterList = this.campaignCryptoList;
+        console.log({ filterList: this.filterList });
+      });
+    }
+  }
+  openModal(content: any) {
+    this.modalService.open(content);
+  }
+  closeModal(content: any) {
+    this.modalService.dismissAll(content);
+  }
+  selectToken(content: any) {
+    if (this.router.url.startsWith('/campaign')) {
+      this.openModal(content);
+    }
+    // get the list of crypto for user
+    this.getUserCrypto();
+
+    // this.walletFacade
+    //   .getAllWallet()
+    //   // .pipe(filter((crypto) => crypto.network === this.selectedNetworkValue))
+    //   .subscribe((res) => {
+    //     console.log(res, 'futur token List');
+    //   });
   }
   //get list of crypto for user
   getusercrypto() {
@@ -259,7 +397,29 @@ export class DropdownCryptoNetworkComponent
     this.selectedCrypto.emit(crypto);
     this.cdref.detectChanges();
   }
+  onCryptoSelected(crypto: any) {
+    // const networkSelectedToken: any = crypto.value.networkSupported?.find(
+    //   (e: any) => {
+    //     e.platform.name.includes(this.selectedCryptoDetails.network);
+    //   }
+    // );
+    this.selectedToken = crypto;
+    // this.selectedCrypto = {
+    //   addr: networkSelectedToken.contract_address,
+    //   name: crypto.value.name,
+    //   type: this.selectedCryptoDetails.network
+    // };
+    this.closeTokenModal(this.tokenModal);
+    console.log(this.selectedToken, 'tokennnSelectedd');
+    // console.log(
+    //   crypto.value.networkSupported?.find((e: any) =>
+    //     e.platform.name.includes(this.selectedCryptoDetails.network)
+    //   ),
+    //   'crypto filtered'
+    // );
 
+    // ?.platform.coin.name
+  }
   selectNetworkValue(network: string) {
     this.firstEmit = true;
     if (this.isCryptoRouter) {
@@ -304,7 +464,7 @@ export class DropdownCryptoNetworkComponent
         this.selectedCrypto.emit(crypto);
       }
     });
-    this.cdref.detectChanges(); 
+    this.cdref.detectChanges();
   }
   ngOnDestroy() {
     this.onDestoy$.next('');
